+++ Mradul-Rathore-Food-App-Case-Study/app/controllers/user.controller.js
@@ -1,4 +1,4 @@
-const db = require("../models")
+const db = require("../config/db.config")
 
 const User = db.users;
 const bcrypt = require("bcryptjs")
